feat(ny): initial add of ny
This may or may not be usable - support will be added to public build machines and server-data in the coming days.

When sharing binaries, please keep https://fivem.net/terms in mind. As to 32-bit CEF, a normal M83 build from Spotify's CEF builders will work.

+++ code/components/citizen-resources-client/src/KVScriptFunctions.cpp
@@ -375,7 +375,7 @@ struct KvpBulkStream
 
 	size_t ReadBulk(uint64_t ptr, void* outBuffer, size_t size)
 	{
-		size_t toRead = std::min(m_value.size() - ptr, size);
+		size_t toRead = std::min(size_t(m_value.size() - ptr), size);
 		memcpy(outBuffer, m_value.data() + ptr, toRead);
 
 		return toRead;
